Helm package should only happen on the master branch
This is because it is very painful to require each pull request to be on
a protected branch. It also means that we require each pull request to
setup a deploy ssh key as well as a gh-pages branch.
In the new model, we would only run that stage on the master branch of
the GPU Operator project.

Signed-off-by: Renaud Gaubert <rgaubert@nvidia.com>