Fix resource creation flow
When resources are created, the object that is given to them is updated.
This leads to the resource array reflecting the kubernetes objects.
When we try to re-create the resource later, we do not get the correct
error because the spec that we try to persist is already filled with
Kubernetes metadata which it should not have.

Signed-off-by: Renaud Gaubert <rgaubert@nvidia.com>