Disable kernel nodeSelector
We currently don't pull the image that is specific to the kernel we are
running on. We are however preventing our driver container to run on
multiple nodes with different kernel versions.

Given that kernel updates are actually a quite frequent phenomenon, we
might as well disable this node selector until we actually have
implemented a strategy for handling matching kernel modules.

Note that the driver container falls back to recompilation if the kernel
module it shipped with don't match the kernel version.

Signed-off-by: Renaud Gaubert <rgaubert@nvidia.com>