enablerepo and releasever injection
this is a re-do of MR 116.

By request of @shivamerla due some changes being lost after a reverse to
fix helm lints

Again:

This patch injects repo and releasever  to driver pod to enable better
handling of rpm dependencies.
This will allow us to config the driver pod more dinamically

Signed-off-by: Carlos Eduardo Arango Gutierrez <carangog@redhat.com>