Revert GFD ordering to start after driver
GFD was started before driver in 1.5.0 release to generate vGPU labels to decide
driver versions to install. But later that logic was moved to driver container itself
and labels are just for tracking vGPU nodes and versions. This change reverts the order
so that delay with label generation can be avoided.

Signed-off-by: Shiva Krishna, Merla <smerla@nvidia.com>