Fix lint steps
This change fixes the lint steps in the CI by ensuring that:
* The tools are installed in the global GOPATH by explicitly disabling GO111MODULE
* The make targets are invoked directly
* The make targets are updated to invoke the commands correctly

Note that the lint target was refering to gpu-feature-discovery and not the
gpu-operator and due to the use of go modules, the gopath manipulation should
not be required and running the lint target yields a number of lint errors.
These could be addressed in a seprate cleanup.

Signed-off-by: Evan Lezar <elezar@nvidia.com>