validator: add new 'metrics' component
This PR extends the `validator` controller to add a new 'metrics'
component.

This components exposes Prometheus metrics on a given HTTP
port (controlled by `METRICS_PORT` environment variable or
`--metrics-port` command-line flag).

The metrics contain various information about the deployment of the
GPU Operator stack on the local node. Prometheus metrics are
self-described (the format requires and exposes a type and help
message).