Cluster policy bundle manifest fixed to contain only one YAML document
When OpenShift's OLM loads a bundle (using the `opm` binary) it scans
the bundle image for manifest files. It then loads the first YAML document
from each of those yaml files. An empty line was added at the beginning
of the cluster policy in a recent commit, right before `---`. This caused
opm to load the manifest as an empty object because that empty line was
enough to count as a document, and since it was the first, it was the only
one used