Always create gpu-clients ConfigMap for MIG manager
This fixes a regression introduced in !326. The updated mig-manager
always expects a gpu-clients file as input, even if the nvidia driver
is deployed as a container. This commit addresses this issue by always
creating the default-gpu-clients ConfigMap even if it does not end up
being used.

Even though the only current use case for the gpu-clients file is
with preinstalled drivers, in the future we plan to add a section
to the gpu-clients config file to specify non-operator managed
daemonsets that need to be shutdown across a MIG reconfiguration.
In this scenario, the gpu-clients config file will be needed
irrespective of how the driver is deployed.

This commit also updates nvidia-mig-manager's SecurityContextConstraint
to AllowHostIPC when drivers are preinstalled.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>