Add initial sandboxed environment support
This commit adds a global sandboxedEnvironments.enabled field in the spec
for enabling sandbox functionality. The aim of this functionality is to
automate the deployment and management of additional software components
needed to provision GPUs for use by Virtual Machines. We use the term
"sandboxed environments" to decouple this functionality from any specific
virtualization solution.

This new functionality will deploy a different set of operands on
different nodes depending on the workload configured for a node via the
node label "nvidia.com/gpu.workload.config". Both GPU-accelerated
containers and VMs can co-exist in the same K8s cluster, but not
on the same worker node with these changes. Mixed deployment scenarios
on a single node with multiple GPUs is left as future work.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>