Add vfio-pci, vgpu-manager, and vgpu-devices validation to the validator component
The added validations are no-ops if the gpu workload configured for the
node does not require the component.

Currently, the vfio-pci validation has not been implemented.
The vGPU validation is simply checking for the presence of a status
file that must be created by another component once all necessary
vGPU devices have been created on the host system.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>