Revert hopper specific MIG configurations in configmap
We did all of our early testing on a hoppper machine that had 8 GPCs
available. Apparently this is NOT the configuration that actual hopper
chips will have when they are shippped though. They will look similar to
an A100-80GB in terms of the MIG profiles they have available.

This change reverts the hopppper specific changes that were previously
made to reflect that.

Signed-off-by: Kevin Klues <kklues@nvidia.com>