Workaround to allow toolkit container to start on recent containerds
Starting with containerd 1.6.9 we started seeing the toolkit container
enter a crashloop whereby it would receive a SIGTERM shortly after
reatarting containerd.

We ended up narrowing the issue down to this commit in contaienerd:
https://github.com/containerd/containerd/commit/1f29fac48ee356918663a4aa1f9880de3e2d6f1a

As part of the debug process, I also noted that the SIGTERM was
originating from an intstance of runc forked by the sandbox process for
the toolkit -- *not* sent directy by containerd or the kubelet.

After investigating for a while what could be causing the issue, it
occurred to me that the commit referenced above mostly just moves around
a bunch of code for the sandbox container. Specifically, it moves a
bunch of code to run *after* the underlying container it is providing a
sandbox for has started running.

So if we restart containerd from within our container, and our
sandbox is not yet done communicating with containerd, it's possible
that we kill containerd in the middle of some operation our sandbox is
performing. Which would then in turn, kill us.

To test this, I added a simple 'sleep 5' just before entrypoint of the
container and viola, issue resolved.

The solution is not very asthetically pleasing, but it at least makes
logical sense, so I think it is an OK solution for now. Hopefully we can
convince containerd to put a more robust solution in place whereby the
container won't have an opportunity to interfere like this in the
future.

Signed-off-by: Kevin Klues <kklues@nvidia.com>