Set custom envvars for driver-validation and toolkit-validation initContainers that are not in the validator pod.
In particular, this allows users to disable the /dev/char symlink creation completely in the operator
by setting "DISABLE_DEV_CHAR_SYMLINK_CREATION=true" in validator.driver.env.
The /dev/char symlinks are created in the driver-validation initContainer, which runs both in the
container-toolkit pod and the operator-validator pod. Without passing custom envvars to the container-toolkit
pod, we will attempt to create /dev/char symlinks even if the user has disabled this.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>