Update the upgrade-controller to also watch for events from NVIDIA driver daemonsets owned by the NVIDIADriver controller.
The upgrade-controller will manage the lifecycle of NVIDIA drivers
regardless if they are owned by the ClusterPolicy or NVIDIADriver controller.
We use a common label selector to identify NVIDIA driver daemonsets.
Additionally, we check the ownerRef field to ensure the source of
the event (the DaemonSet object) is in fact managed by the GPU Operator.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>