Do not update k8s-driver-manager image version if other driver spec has not changed
This avoids undesired driver upgrades when the operator is upgraded. When the operator
is upgraded, the image environment variables set in the operator image will likely
have been bumped. The new values of these environment variables, which act as default
values for operand images, should only be applied to the driver DaemonSet iff the driver
spec has been updated, and thus, the driver upgrade is required.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>