Create driver-ready file with content and simplify operand startup scripts
This commit updates the driver validation to always create a 'driver-ready' file,
regardless if the driver is installed on the host or not. It also populates this file
with a list of environment variables, one per line, which are required by some operands.

The startup scripts for several operands are simplified to simply source the content
of this file before executing the main program for the container.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>