Only chroot when validating a host installed driver
This commit updates our driver validator code to only chroot when
validating a host installed driver. When validating a driver container
install, we discover the paths to libnvidia-ml.so.1 and nvidia-smi at
the driver container root and then run 'LD_PRELOAD=/driverRoot/path/to/libnvidia-ml.so.1 nvidia-smi'.

This sets the stage for validating driver container installs where driverRoot
does not represent a full filesystem hiearchy that one can 'chroot' into.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>