Add a 'hostPaths.rootFS' field to ClusterPolicy
RootFS represents the path to the root filesystem of the host.
This is used by components that need to interact with the host filesystem
and as such this must be a chroot-able filesystem.
Examples include the MIG Manager and Toolkit Container which may need to
stop, start, or restart systemd services.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>
Co-authored-by: Angelos Kolaitis <neoaggelos@gmail.com>