Always mount a driver container install to '/driver-root' in our operand containers except the driver-validator
Having a static path inside our containers will make it easier when driverRoot is a configurable field.
If driverRoot is set to a custom path, we can transform the host path for the volume while keeping
the container path unchanged.

The driver-validation initContainer is the exception to this rule. From the driver validation
initContainer, the container path must match the host path otherwise the /dev/char symlinks will not
resolve correctly on the host. The target of the symlinks must correspond to the path of the device
nodes on the host. For example, when the NVIDIA device nodes are present under `/run/nvidia/driver/dev`
on the host, running the following command from inside the container would create an invalid symlink:

  ln -s /driver-root/dev/nvidiactl /host-dev-char/195:255

while running the below command from inside the container would create a valid symlink:

  ln -s /run/nvidia/driver/dev/nvidiactl /host-dev-char/195:255

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>