Detect if container runtime versions on worker nodes support CDI
This commit adds a 'runtimeSupportsCDI' field to the ClusterPolicyController struct.
When fetching the container runtime version strings from worker nodes, the 'runtimeSupportsCDI'
field is set accordingly. If any worker node is running containerd < 1.7.0, then we
set runtimeSupportsCDI=false.

As part of this commit, we now return an error if there are different container runtimes
running on different worker nodes.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>