Set runtimeClass for operands unless CDI is disabled and cri-o is the runtime
When CDI is disabled and cri-o is the container runtime, we fallback to
installing the prestart OCI hook. Our operands will depend on the prestart
hook for getting access to GPUs.

In all other scenarios, we want our operands to leverage the 'nvidia'
runtime class to get access to GPUs.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>