Update github issue template for raising bug reports
Currently, when attempting to create a GitHub issue against this repository
no issue template is used or presented to the user.

This commit updates the content of the bug report template. This commit
also migrates us from the legacy workflow to the new workflow for managing
GitHub issue templates -- the main change is that template file(s) are
now to be stored in the .github/ISSUE_TEMPLATE directory.

See https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/about-issue-and-pull-request-templates

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>