[ci] Bump containerd version to 1.7.27 in holodeck config
Containerd 1.7.0 is the first release that supports CDI. CDI devices were exclusively
passed to containerd via annotations.
https://github.com/containerd/containerd/releases/tag/v1.7.0

In containerd 1.7.2 support was added for making use
of the new CRI fields 'Config.CDIDevices' so that CDI devices could be passed through the CRI
instead of annotations.
https://github.com/containerd/containerd/releases/tag/v1.7.2

This commit updates our holodeck configuration to a version of containerd that supports CDI
and the new CRI fields. containerd 1.7.27 is chosen since it is the latest release on the 1.7
branch that is available in the Ubuntu 22.04 repositories.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>