[dcgm-exporter] volume mount /run/nvidia if using hpc job files
dcgm-exporter reads files from a user-defined directory to determine
gpu-to-job mappings when hpc job mapping is enabled. Currently gpu-operator
will only volume mount /run/nvidia for initialization, but not the long
running dcgm-exporter pod, so it is not possible to configure a job mapping
directory that maps back to the host.

This patch mounts /run/nvidia as readonly into the dcgm-exporter pod if hpc
job mapping is enabled. This lets hpc job mappings be communicated from the
host file system to dcgm-exporter via files in a user-defined directory
under /run/nvidia.

Signed-off-by: Anthony Romano <aromano@nvidia.com>