Add ld.so.cache file to the gpu-operator container image
The distroless image does not contain the ld.so.cache file. The update-ldcache
createContainer hook will skip updating the container's ldcache if this file
does not exist. This can lead to issues running the CUDA vectorAdd sample
if the NVIDIA libraries are not present in the default dynamic linker search
path(s).

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>