Allow config sources to be specified for containerd
By default the toolkit container extracts the current containerd config by running
the `containerd config dump` command and if that fails we read the existing config file.

This change allows ordering of these sources to be defined via the RUNTIME_CONFIG_SOURCE
environment variable. This further allows an explicit path to be specified for the file
config source so that there is a separation between the config file being modified by the
toolkit container and the config used as the source for drop-in files, for example.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>