Revert back to using the hook on cri-o when CDI is disabled
When CDI is not enabled and cri-o is the container runtime,
we continue to install the OCI prestart hook as opposed to adding nvidia
runtime handlers to the cri-o configuration. Users can override this behavior
and have nvidia runtime handlers added to the cri-o configuration by setting
the 'CRIO_CONFIG_MODE' environment variable to 'config' in the toolkit container.
However, one should note setting 'CRIO_CONFIG_MODE' to 'config' in this case
(when CDI is not enabled) would result in the 'nvidia' runtime being set as
the default runtime. While this should work in theory, it is a significant
change -- which was the primary motivation to continue using the OCI prestart
hook by default in this case.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>