Run nvidia-smi after modules are loaded in driver ds startup probe
This commit eliminates the race condition where the startup probe
in the driver daemonset runs after the kernel modules are built
(and installed) but before the modules are loaded into the kernel.
In this case, the invocation of nvidia-smi (by the startup probe)
is what is actually loading the nvidia kernel module and not the
modprobe we perform in our driver container scripts. As a result,
the nvidia driver will be loaded with a default configuration --
none of the custom kernel module parameters provided by users
(via a configmap) or set by our driver container will get applied.

Signed-off-by: Christopher Desiniotis <cdesiniotis@nvidia.com>
(cherry picked from commit 90e757a5699e917bf9aaac0b29e1b46affa5eb30)